Add CMake rules to generate and build zxdg-unstable-shell-v6 protocol
This is quite complicated since it will be #included in
WinSystemWayland.h, which is included by many files, meaning that
the wayland-extra-protocols.hpp header must be generated before
compiling any Kodi source. This is guaranteed by adding an extra
CMake target that is included as dependency in the main Kodi lib file
and all subdirectory static libs.