Do not change preferred monitor when it is not available
When the monitor set by the user is not available, not including it
in the list of settable monitors would lead to a false monitor switch
including an annoying dialog box each time the system settings window
is entered. So just include it in the list until the user selects
something else. This also means that the setting will be preserved so we
can (threotically) switch to it on hot-plug events.