[vaapi] Set export surface hint on creation where appropriate
Mesa can avoid allocating field-layout buffers that then have to be
woven on export this way.
Also, some decoders such as HEVC on AMD do not support the field
layout and will reallocate the surfaces anyway, leading to problems
with post-processing when the post-processing surfaces are in field
layout (mesa cannot process this case). This situation can be avoided
by making sure the postproc output surfaces are always in frame
layout.